Bump Version 1.3.91

+++ builds/ansible/ddb-ansible.py
@@ -129,7 +129,7 @@ def run_module():
 # File   : ./source/ddb/version.py
 # ############################################################################
 
-__version__='1.3.90'
+__version__='1.3.91'
 
         
 # ############################################################################
@@ -4439,6 +4439,9 @@ def swap_files(path, temp,key_uuid):
     norm_path=normalize_path(path)
     if os.path.exists(norm_path)==True:
         remove_temp_file(norm_path)
+    if os.path.exists(norm_path)==True:
+        lock.error("Lock","MASTER FILE WONT DELETE")
+        exit(1)
     if lock.debug: lock.info("Lock","Copying temp to master {0} <- {1}".format(norm_path,temp))
     lock.copy_file(temp, norm_path)
     time.sleep(.001)
